pyproject.toml
[project]
name = "paisalog"
version = "0.1.0"
description = "A personal expense and income tracking application"
requires-python = ">=3.10"
dependencies = [
    "django>=5.0.0",
    "django-environ>=0.11.2",
    "django-widget-tweaks>=1.5.0",
    "django-tables2>=2.7.0",
    "django-filter>=24.1",
    "django-bootstrap5>=23.4",
    "django-menu-generator>=1.1.0",
    "tablib>=3.8.0",
    "openpyxl>=3.1.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["paisalog", "core"]

[tool.ruff]
line-length = 120
target-version = "py310"
select = ["E", "F", "B", "I"]
ignore = []

[tool.ruff.isort]
known-first-party = ["paisalog", "core"] 
